Remove old `static` filter library
The ipfw-named filter has statically query schema for packet attributes.
New filters allow one to customize attribute query and lookup order.
