+++ modules/acl/api/meson.build
@@ -1,5 +1,7 @@
 cp_dependencies = [
   lib_common_dep,
+  lib_filter_dep,
+  lib_packet_dp_dep,
   lib_config_cp_dep,
   lib_counters_dep,
 ]
